updated with client-sdk & added dynamoose deps
