+++ scene-store-service/package.json
@@ -28,6 +28,7 @@
         "aws-lambda": "^1.0.6",
         "aws-sdk": "^2.783.0",
         "aws-serverless-express": "^3.3.8",
+        "dynamoose": "^2.4.0",
         "nanoid": "^3.1.16",
         "reflect-metadata": "^0.1.13",
         "rimraf": "^3.0.2",
@@ -51,9 +52,9 @@
         "jest": "26.4.2",
         "plugin": "^0.3.3",
         "prettier": "^1.19.1",
+        "serverless-domain-manager": "^5.1.0",
         "serverless-dotenv-plugin": "^3.1.0",
         "serverless-dynamodb-local": "^0.2.39",
-        "serverless-domain-manager": "^5.1.0",
         "serverless-offline": "^6.8.0",
         "serverless-plugin-optimize": "^4.1.4-rc.1",
         "supertest": "^4.0.2",
