+++ scene-store-service/src/app.service.ts
@@ -1,47 +1,28 @@
 import { Injectable } from '@nestjs/common';
 import * as AWS from "aws-sdk"
 import { DocumentClient } from 'aws-sdk/lib/dynamodb/document_client';
+import * as dynamoose from "dynamoose";
 
-const dynamoDb = new AWS.DynamoDB.DocumentClient();
-
-interface DynamoSceneTable {
-  id: string
-  cachedPreview: string
-  fileId: string
-  name: string
-  path: string
-  route: string
-  layersId: string[]
-  componentsId: string[]
-  type: "screen" | "component"
-  width: number
-  height: number
-}
-
-// 1 - tree
-
-interface Layer {
-  id: string
-  key: string
-  cachedPreview: string
-  parentSceneId: string
-  thisSceneId?: string
-  width: number
-  height: number
-  index: number
-  node: object
-}
-
+const dynamodb = dynamoose.aws.ddb()
+const TABLE = process.env.DYNAMODB_TABLE
 
 @Injectable()
 export class AppService {
-  getHello(): string {
+  async getHello(): Promise<string> {
     return 'Welcome to bridged hosting service. Learn more at https://github.com/bridgedxyz/services/';
   }
 
 
-  putScene() {
+  async putScene() {
+    // test
 
+    const User = dynamoose.model(TABLE, { "id": Number, "name": String });
+    const myUser = new User({
+      "id": 1,
+      "name": "Tim"
+    });
+    const saved = await myUser.save()
+    console.log('saved', saved)
   }
 
   putLayer() {
